+++ classes/class-apis.php
@@ -23,11 +23,12 @@ class Apis {
 	private static ?Apis $instance = null;
 
 	/**
-	 * Constructor for Init-Handler.
+	 * Constructor for this object.
 	 */
 	private function __construct() {
 		// hooks.
 		add_action( 'admin_action_easy_language_export_api_log', array( $this, 'export_api_log' ) );
+		add_action( 'admin_action_easy_language_clear_api_log', array( $this, 'clear_api_log' ) );
 	}
 
 	/**
@@ -166,6 +167,47 @@ class Apis {
 		foreach ( $entries as $row ) {
 			fputcsv( $fp, $row );
 		}
+		exit;
+	}
+
+	/**
+	 * Clear API log for specific API.
+	 *
+	 * @return void
+	 * @noinspection PhpNoReturnAttributeCanBeAddedInspection
+	 */
+	public function clear_api_log(): void {
+		// check nonce.
+		check_admin_referer( 'easy-language-clear-api-log', 'nonce' );
+
+		// get name of the api to clear.
+		$clear_api = isset( $_GET['api'] ) ? sanitize_text_field( wp_unslash( $_GET['api'] ) ) : '';
+
+		// bail if no api is given.
+		if ( empty( $clear_api ) ) {
+			// redirect user back.
+			wp_safe_redirect( wp_get_referer() );
+			exit;
+		}
+
+		// get api object.
+		$api_object = $this->get_api_by_name( $clear_api );
+
+		// bail if no API object could be loaded.
+		if ( ! $api_object instanceof Base ) {
+			// redirect user back.
+			wp_safe_redirect( wp_get_referer() );
+			exit;
+		}
+
+		// get db object.
+		global $wpdb;
+
+		// delete the entries.
+		$wpdb->query( $wpdb->prepare( 'DELETE FROM ' . Log_Api::get_instance()->get_table_name() . ' WHERE `api` = %s', array( $api_object->get_name() ) ) );
+
+		// redirect user back.
+		wp_safe_redirect( wp_get_referer() );
 		exit;
 	}
 }