+++ classes/class-log-table.php
@@ -174,4 +174,33 @@ class Log_Table extends WP_List_Table {
 		// return the HTML-code for the icon of this status.
 		return $list[ $status ];
 	}
+
+	/**
+	 * Add export-buttons on top of table.
+	 *
+	 * @param string $which Position.
+	 * @return void
+	 */
+	public function extra_tablenav( $which ): void {
+		// bail if this is not for top view.
+		if ( 'top' !== $which ) {
+			return;
+		}
+
+		if ( ! empty( $this->items ) ) {
+			// define clear URL.
+			$url = add_query_arg(
+				array(
+					'action' => 'easy_language_clear_log',
+					'nonce'  => wp_create_nonce( 'easy-language-clear-log' )
+				),
+				get_admin_url() . 'admin.php'
+			);
+			?><a href="<?php echo esc_url( $url ); ?>" class="button"><?php echo esc_html__( 'Clear log', 'easy-language' ); ?></a><?php
+		} else {
+			?>
+			<span class="button disabled" title="<?php echo esc_html__( 'Choose an API above to clear the log', 'easy-language' ); ?>"><?php echo esc_html__( 'Clear log', 'easy-language' ); ?></span>
+			<?php
+		}
+	}
 }