+++ changelog.md
@@ -5,6 +5,8 @@
 ### Added
 
 - Added option for embolden negative words during simplification via SUMM AI
+- Added option to clear API log
+- Added option to clear the plugin log
 
 ### Optimized